requirements.txt
streamlit==1.20.0
PyPDF2==1.26.0
pandas==1.5.3
python-pptx==0.6.21
python-docx==0.8.11
requests==2.28.1
beautifulsoup4==4.11.1
langchain==0.0.200  # Adjust this based on the latest version
langchain-google-genai==0.0.1  # Adjust this based on the latest version
langchain-community==0.0.1  # Adjust this based on the latest version
python-dotenv==0.20.0